[SPARK-46264][PS][CONNECT][TESTS] Re-organize the resampling tests
### What changes were proposed in this pull request?
Re-organize the resampling tests

### Why are the changes needed?
Re-organize the resampling tests by topics to be more consistent with [Pandas tests](https://github.com/pandas-dev/pandas/tree/main/pandas/tests)

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44180 from zhengruifeng/reorg_test_resample.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
