+++ dev/sparktestsupport/modules.py
@@ -740,7 +740,10 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.test_ops_on_diff_frames_groupby_expanding",
         "pyspark.pandas.tests.test_ops_on_diff_frames_groupby_rolling",
         "pyspark.pandas.tests.test_repr",
-        "pyspark.pandas.tests.test_resample",
+        "pyspark.pandas.tests.resample.test_on",
+        "pyspark.pandas.tests.resample.test_error",
+        "pyspark.pandas.tests.resample.test_missing",
+        "pyspark.pandas.tests.resample.test_timezone",
         "pyspark.pandas.tests.test_frame_resample",
         "pyspark.pandas.tests.test_series_resample",
         "pyspark.pandas.tests.test_reshape",
@@ -984,7 +987,10 @@ pyspark_pandas_connect_part0 = Module(
         "pyspark.pandas.tests.connect.test_parity_namespace",
         "pyspark.pandas.tests.connect.test_parity_numpy_compat",
         "pyspark.pandas.tests.connect.test_parity_repr",
-        "pyspark.pandas.tests.connect.test_parity_resample",
+        "pyspark.pandas.tests.connect.resample.test_parity_error",
+        "pyspark.pandas.tests.connect.resample.test_parity_missing",
+        "pyspark.pandas.tests.connect.resample.test_parity_on",
+        "pyspark.pandas.tests.connect.resample.test_parity_timezone",
         "pyspark.pandas.tests.connect.test_parity_scalars",
         "pyspark.pandas.tests.connect.test_parity_series_conversion",
         "pyspark.pandas.tests.connect.test_parity_series_datetime",
