+++ python/pyspark/pandas/tests/connect/resample/test_parity_timezone.py
@@ -16,25 +16,19 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_resample import ResampleTestsMixin, ResampleWithTimezoneMixin
+from pyspark.pandas.tests.resample.test_timezone import ResampleTimezoneMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
-from pyspark.testing.pandasutils import PandasOnSparkTestUtils, TestUtils
+from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
-class ResampleParityTests(
-    ResampleTestsMixin, PandasOnSparkTestUtils, TestUtils, ReusedConnectTestCase
-):
-    pass
-
-
-class ResampleWithTimezoneTests(
-    ResampleWithTimezoneMixin, PandasOnSparkTestUtils, TestUtils, ReusedConnectTestCase
+class ResampleParityTimezoneTests(
+    ResampleTimezoneMixin, PandasOnSparkTestUtils, ReusedConnectTestCase
 ):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_resample import *  # noqa: F401
+    from pyspark.pandas.tests.connect.resample.test_parity_timezone import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
